+++ examples/saas-dynamic-dashboards/frontend/app/Client/client.ts
@@ -1,0 +1,10 @@
+import axios, { AxiosRequestConfig } from "axios";
+
+export const client = async (config: AxiosRequestConfig) => {
+    try {
+        const res = await axios(config)
+        return res.data
+    } catch (error) {
+        throw error
+    }
+}