+++ sdk-python/copilotkit/langgraph_agent.py
@@ -157,7 +157,7 @@ class LangGraphAgent(Agent):
         if agent is not None:
             logger.warning("Warning: agent is deprecated, use graph instead")
 
-        if merge_state is None:
+        if merge_state is not None:
             logger.warning("Warning: merge_state is deprecated, use copilotkit_config instead")
 
         if graph is None and agent is None:
@@ -261,7 +261,7 @@ class LangGraphAgent(Agent):
         ):
 
         default_config = ensure_config(cast(Any, self.langgraph_config.copy()) if self.langgraph_config else {}) # pylint: disable=line-too-long
-        config = {**default_config, **(self.graph.config or {}), **config}
+        config = {**default_config, **(self.graph.config or {}), **(config or {})}
         config["configurable"] = {**config.get("configurable", {}), **(config["configurable"] or {})}
         config["configurable"]["thread_id"] = thread_id