+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,79 @@
 # ui
 
+## 1.8.15-next.0
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.15-next.0
+- @copilotkit/shared@1.8.15-next.0
+
+## 1.8.14
+
+### Patch Changes
+
+- 9cf1fda: - fix append follow-up when actions disable followUp
+  - Create stupid-nails-travel.md
+  - fixup
+- 9cf1fda: - fix append follow-up when actions disable followUp
+- Updated dependencies [34a78d8]
+  - @copilotkit/shared@1.8.14
+  - @copilotkit/runtime-client-gql@1.8.14
+
+## 1.8.14-next.5
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.14-next.5
+- @copilotkit/shared@1.8.14-next.5
+
+## 1.8.14-next.4
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.14-next.4
+- @copilotkit/shared@1.8.14-next.4
+
+## 1.8.14-next.3
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.14-next.3
+- @copilotkit/shared@1.8.14-next.3
+
+## 1.8.14-next.2
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.14-next.2
+- @copilotkit/shared@1.8.14-next.2
+
+## 1.8.14-next.1
+
+### Patch Changes
+
+- Updated dependencies [34a78d8]
+  - @copilotkit/shared@1.8.14-next.1
+  - @copilotkit/runtime-client-gql@1.8.14-next.1
+
+## 1.8.14-next.0
+
+### Patch Changes
+
+- 9cf1fda: - fix append follow-up when actions disable followUp
+  - Create stupid-nails-travel.md
+  - fixup
+- 9cf1fda: - fix append follow-up when actions disable followUp
+  - @copilotkit/runtime-client-gql@1.8.14-next.0
+  - @copilotkit/shared@1.8.14-next.0
+
+## 1.8.13
+
+### Patch Changes
+
+- 7fcf5c4: - fix followUp check
+  - @copilotkit/runtime-client-gql@1.8.13
+  - @copilotkit/shared@1.8.13
+
 ## 1.8.13-next.3
 
 ### Patch Changes