+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,84 @@
 # ui
 
+## 1.8.15-next.0
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.15-next.0
+- @copilotkit/react-core@1.8.15-next.0
+- @copilotkit/shared@1.8.15-next.0
+
+## 1.8.14
+
+### Patch Changes
+
+- Updated dependencies [34a78d8]
+- Updated dependencies [9cf1fda]
+- Updated dependencies [9cf1fda]
+  - @copilotkit/shared@1.8.14
+  - @copilotkit/react-core@1.8.14
+  - @copilotkit/runtime-client-gql@1.8.14
+
+## 1.8.14-next.5
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.14-next.5
+- @copilotkit/react-core@1.8.14-next.5
+- @copilotkit/shared@1.8.14-next.5
+
+## 1.8.14-next.4
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.14-next.4
+- @copilotkit/react-core@1.8.14-next.4
+- @copilotkit/shared@1.8.14-next.4
+
+## 1.8.14-next.3
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.14-next.3
+- @copilotkit/react-core@1.8.14-next.3
+- @copilotkit/shared@1.8.14-next.3
+
+## 1.8.14-next.2
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.14-next.2
+- @copilotkit/react-core@1.8.14-next.2
+- @copilotkit/shared@1.8.14-next.2
+
+## 1.8.14-next.1
+
+### Patch Changes
+
+- Updated dependencies [34a78d8]
+  - @copilotkit/shared@1.8.14-next.1
+  - @copilotkit/react-core@1.8.14-next.1
+  - @copilotkit/runtime-client-gql@1.8.14-next.1
+
+## 1.8.14-next.0
+
+### Patch Changes
+
+- Updated dependencies [9cf1fda]
+- Updated dependencies [9cf1fda]
+  - @copilotkit/react-core@1.8.14-next.0
+  - @copilotkit/runtime-client-gql@1.8.14-next.0
+  - @copilotkit/shared@1.8.14-next.0
+
+## 1.8.13
+
+### Patch Changes
+
+- Updated dependencies [7fcf5c4]
+  - @copilotkit/react-core@1.8.13
+  - @copilotkit/runtime-client-gql@1.8.13
+  - @copilotkit/shared@1.8.13
+
 ## 1.8.13-next.3
 
 ### Patch Changes